# Speed up trade jump generation