--- a/PyRoute/Calculation/TradeCalculation.py
+++ b/PyRoute/Calculation/TradeCalculation.py
@@ -150,10 +150,14 @@ class TradeCalculation(RouteCalculation):
 
     def _raw_ranges(self):
         max_route_dist = max(self.btn_range)
+        max_range = self.galaxy.max_jump_range
+        min_btn = self.min_btn
 
         ranges = [(star, neighbour) for (star, neighbour) in itertools.combinations(self.galaxy.ranges, 2)
                   if not star.is_redzone and not neighbour.is_redzone
-                  and star.distance(neighbour) <= self._max_dist(star.wtn, neighbour.wtn, True)]
+                  and (dist := star.distance(neighbour)) <= self._max_dist(star.wtn, neighbour.wtn, True)
+                  and self._get_btn_upper_bound(star, neighbour, max_range, min_btn, distance=dist) >= min_btn
+                  ]
         self.logger.info("Routes with endpoints more than " + str(max_route_dist) + " pc apart, trimmed")
 
         return ranges