--- a/Tests/Pathfinding/testApproximateShortestPathForest.py
+++ b/Tests/Pathfinding/testApproximateShortestPathForest.py
@@ -128,7 +128,6 @@ class testApproximateShortestPathForest(baseTest):
                 exp_dist = expected_string[str(rawstar)]
             expected_distances[item] = exp_dist
 
-        expected_distances[19] = float('+inf')
         distance_check = list(expected_distances.values()) == approx.distances[:, 0]
         self.assertTrue(distance_check.all(), "Unexpected distances after SPT creation")