--- a/Tests/Pathfinding/testApproximateShortestPathTree.py
+++ b/Tests/Pathfinding/testApproximateShortestPathTree.py
@@ -80,7 +80,7 @@ class testApproximateShortestPathTree(baseTest):
         src = stars[20]
         targ = stars[19]
 
-        expected = 0.0
+        expected = 0.833
         actual = approx.lower_bound(src, targ)
         self.assertAlmostEqual(expected, actual, 3, "Unexpected lower bound value")
 
@@ -141,7 +141,7 @@ class testApproximateShortestPathTree(baseTest):
         result = approx.lower_bound_bulk(target)
         self.assertIsNotNone(result)
         result = result[active_nodes]
-        expected = np.array([15.833, float('+inf'), 0])
+        expected = np.array([15.833, 15, 0])
         np.testing.assert_array_almost_equal(expected, result, 0.000001, "Unexpected bounds array")
 
     def test_drop_first_level_intermediate_nodes_in_same_component(self):
@@ -258,7 +258,6 @@ class testApproximateShortestPathTree(baseTest):
             if str(rawstar) in expected_string:
                 exp_dist = expected_string[str(rawstar)]
             expected_distances[item] = exp_dist
-        expected_distances[19] = float('+inf')
 
         distance_check = list(expected_distances.values()) == approx.distances[:, 0]
         self.assertTrue(distance_check.all(), "Unexpected distances after SPT creation")
@@ -318,7 +317,6 @@ class testApproximateShortestPathTree(baseTest):
             if str(rawstar) in expected_string:
                 exp_dist = expected_string[str(rawstar)]
             expected_distances[item] = exp_dist
-        expected_distances[19] = float('+inf')
 
         distance_check = list(expected_distances.values()) == approx.distances[:, 0]
         self.assertTrue(distance_check.all(), "Unexpected distances after SPT creation")