--- a/Tests/PathfindingFiles/single_source_distances_ibara_subsector_from_0101.json
+++ b/Tests/PathfindingFiles/single_source_distances_ibara_subsector_from_0101.json
@@ -6,7 +6,7 @@
   "Selsinia (Zarushagar 0201)": 54.0, "Arxani (Zarushagar 0209)": 283.0, "Point Zulu (Zarushagar 0302)": 74.0,
   "Makkus (Zarushagar 0304)": 119.0, "Freeport (Zarushagar 0305)": 145.0, "Woden (Zarushagar 0306)": 165.0,
   "Ichiban (Zarushagar 0309)": 255.0, "Shadishi (Zarushagar 0310)": 283.0, "Rogreis (Zarushagar 0402)": 101.0,
-  "Bolten (Zarushagar 0404)": 144.0, "Ulaun (Zarushagar 0405)": 191.0, "Nedadzia (Zarushagar 0701)": 329.0,
+  "Bolten (Zarushagar 0404)": 144.0, "Ulaun (Zarushagar 0405)": 191.0, "Nedadzia (Zarushagar 0701)": 277.0,
   "Ledadzia (Zarushagar 0704)": 239.0, "Dorevann (Zarushagar 0708)": 313.0, "New Orlando (Zarushagar 0710)": 403.0,
-  "Airvae (Zarushagar 0801)": 309.0, "Sufizi (Zarushagar 0802)": 281.0, "Cyrpus (Zarushagar 0803)": 261.0,
+  "Airvae (Zarushagar 0801)": 305.0, "Sufizi (Zarushagar 0802)": 281.0, "Cyrpus (Zarushagar 0803)": 261.0,
   "Gishin (Zarushagar 0804)": 236.0, "Ginshe (Zarushagar 0805)": 263.0, "Madagast (Zarushagar 0806)": 239.0}