Couple of things to speed up trade jump generation a little bit:
1 - Pre-filter raw ranges by an _upper bound_ on what that route's BTN _can_ be - if the upper bound doesn't reach the minimum BTN to calculate a route for, then there's no point calculating the actual BTN, as that won't reach the minimum BTN either.
2 - Break galaxy.ranges into two sets:
    - hiball - stars not redzoned, have minimum **WTN** or better;
    - loball - stars not redzoned, do not have minimum WTN;
    - the third, implicit set is stars that _are_ redzoned, which aren't considered further

Calculate routes as union of filtered results of **itertools.connections(hiball, 2)**, **itertools.connections(loball, 2)** and **itertools.product(hiball, loball)**.  As only the hiball connections have both ends meeting or beating minimum WTN, that's the only group that needs to be filtered by the BTN upper bound - the others just need to check the endpoints are within max jump range.

On full-orchestra, jump generation went from 1257 sec (20 min 57 sec) to 1048 sec (17 min 28 sec), a 16.6% reduction